perf(flow): optimize pkt_ring_entry to 64 bytes (1 cache line)
Change bytes field from int64_t (8 bytes) to uint32_t (4 bytes).
The uint32_t fits in the padding gap after struct flow, eliminating
wasted space and aligning the struct to exactly 64 bytes.

Layout:
- struct flow: 44 bytes (hash key)
- uint32_t bytes: 4 bytes (fits in padding gap)
- struct timeval: 16 bytes (8-byte aligned at offset 48)
- Total: 64 bytes

Max packet size of 4GB is more than sufficient (MTU ~1500, jumbo ~9KB).

Ring buffer memory at 262K entries: 16.8 MB (was 18.9 MB with 72 bytes).

+++ deps/toptalk/flow.h
@@ -367,16 +367,20 @@ struct flow_pkt {
 };
 
 /* Compact ring buffer entry - stores only what's needed for expiration.
- * Used instead of full flow_pkt to reduce memory ~10x (736 -> 68 bytes).
+ * Used instead of full flow_pkt to reduce memory ~10x (736 -> 64 bytes).
  * The ring buffer only needs to track:
  * - flow key for hash table lookup during expiration
  * - bytes to subtract from flow totals
  * - timestamp to determine when entry expires
+ *
+ * Layout optimized for 64-byte cache line:
+ * - bytes (uint32_t) fits in padding gap after flow (44 bytes + 4 = 48)
+ * - timestamp (struct timeval) is 8-byte aligned at offset 48
  */
 struct pkt_ring_entry {
 	struct flow flow;           /* 44 bytes - hash key for lookup */
-	int64_t bytes;              /* 8 bytes - to subtract on expiration */
+	uint32_t bytes;             /* 4 bytes - fits in padding, max 4GB/pkt */
 	struct timeval timestamp;   /* 16 bytes - for age check */
-};  /* Total: 68 bytes */
+};  /* Total: 64 bytes (1 cache line) */
 
 #endif
